IOTA-Bridge implementation of the "Sensor to IOTA-Bridge Repairing" feature
The Repairing allows the sensor to use compressed messages even if a new IOTA-Bridge is used that has not cached the streams-channel-id of the sensor. Repairing is an extension of the application layer protocol that manages this edge case and allows the sensor to hand the channel-id in later. After the IOTA-Bridge receives the missing channel-id, the original compressed request is processed and the resulting response is returned to the sensor.

The new IOTA-Bridge code has not been tested together with the sensor yet.