New property initialization_cnt in SUSEE http protocol allows iota-bridge to detect reinitialized sensors and to update the new channel-id via transmit endpoint
The initialization_cnt is stored in the sensor wallet file and is incremented when the sensor subscribes to a new streams channel.

The iota-bridge stores the initialization_cnt together with the channel-id in its SQLite db to detect reinitialized sensors.

The initialization_cnt is loged to the console by the sensor and by the management console which is informed via the subscription-confirmation about the initialization_cnt.

Sensors can only be reinitialized 255 times. If this limit is reached a warning is loged by the  sensor and by the management console.