Code has been migrated to use latest Streams version from Stardust branch - compilable but not usable
TODO:
Streams for stardust needs an indexer to map Streams message ids to tangle block ids. This will be implemented based on the inx-collector (by Teleconsys).

A docker environment is prepared for development purposes to test the SUSEE POC code together with a private tangle and the inx-collector.

An extended version of the inx-collector (inx-streams-collector) needs to be implemented as a for of the original Teleconsys repository. Further on the inx-streams-collector needs to be integrated into the SUSEE Streams POC code.